+++ CargaUy-ejb/src/main/java/tse/java/service/IVehiculosService.java
@@ -12,14 +12,14 @@ import javax.ejb.Local;
 public interface IVehiculosService {
 
     public List<VehiculoDTO> obtenerVehiculos();
-    public VehiculoDTO obtenerVehiculoPorId(Long id);
+    public VehiculoDTO obtenerVehiculoPorId(int id);
     public void agregarVehiculo(Vehiculo nuevoVehiculo);
     public void modificarVehiculo(VehiculoDTO vehiculoModificado);
-    public void eliminarVehiculo(Long id);
+    public void eliminarVehiculo(int id);
     public VehiculoDTO obtenerVehiculoMatriculaPais(String matricula, String pais);
-    public List<PesajeDTO> listarGuiasDeVehiculo(Long id, Date fecha);
+    public List<PesajeDTO> listarGuiasDeVehiculo(int id, Date fecha);
     public boolean viajeContieneGuia(VehiculoDTO v, GuiaDeViajeDTO g);
-    public void asignarGuia(Long vehiculo_id, AsignacionDTO a);
+    public void asignarGuia(int vehiculo_id, AsignacionDTO a);
     public void borrarGuia(int numero_guia);
     public VehiculoDTO buscarVehiculoPorGuia(int numero);
 }
